test(e2e): house collection specs under the resource shard
Move collection Playwright coverage into tests/e2e/resource so it runs
with the existing resource CI job. Fold cancel-create into
create-collection.test.ts and drop the duplicate permissions file.
