+++ apps/studio/tests/e2e/resource/assign-tags.test.ts
