+++ apps/studio/tests/e2e/resource/collection-display.test.ts
