+++ apps/studio/tests/e2e/resource/collection-table.test.ts
