+++ apps/studio/tests/e2e/resource/create-collection-link.test.ts
