+++ apps/studio/tests/e2e/resource/create-collection-page.test.ts
