+++ apps/studio/tests/e2e/resource/edit-collection-link.test.ts
