+++ apps/studio/tests/e2e/resource/edit-collection-page.test.ts
