+++ apps/studio/tests/e2e/resource/filters.test.ts
