+++ apps/studio/tests/e2e/resource/legacy-collection-experience.test.ts
