+++ apps/studio/tests/e2e/resource/publish-collection.test.ts
